Added BitArray, Voxel Node now uses Bit Arrays
Added the bit. You can set Bits. Get the binary number of all the bits as a string. get the number of all the bits as a byte, short, in, long. Fixed formatting on Bit. Got rid of the Debug.Log for the NodeToSting length, the length variable on the Voxel Grid. Voxel node know uses the BitArray.