Added TMP. Fixed Texture problems read Description
Added Text Mesh pro into the project so I could use quantum console.
Fixed the 3 problems with the textures. 1. unnecessary RAM use by not using the textureColor[], and instead using math to find which place it should go on the texture. 2. Whole voxel not being the same color. I fixed this by insead of putting a new color onto the voxel for each voxel, I used a index for the texture coords then converted it to X Y. Separated the functions for setting the color on the texture, and assigning the UV. 3 color from other pixel showing on the egde. I just made the UV only show the middle of the pixel, and not the whole thing.